+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased][]
 
+## [1.4.0][] - 2019-11-22
+
 ### Added
 
 - Support for additional HTML class namespace
@@ -80,7 +82,8 @@
 - Initial implementation
 
 
-[Unreleased]: https://github.com/niksy/x-autosuggest/compare/v1.3.1...HEAD
+[Unreleased]: https://github.com/niksy/x-autosuggest/compare/v1.4.0...HEAD
+[1.4.0]: https://github.com/niksy/x-autosuggest/compare/v1.3.1...v1.4.0
 [1.3.1]: https://github.com/niksy/x-autosuggest/compare/v1.3.0...v1.3.1
 [1.3.0]: https://github.com/niksy/x-autosuggest/compare/v1.2.2...v1.3.0
 [1.2.2]: https://github.com/niksy/x-autosuggest/compare/v1.2.1...v1.2.2